Version Number Checking
The script now has the function to check the installed version against the latest version that is online. 
If the online version is newer Chromium is updated.
I copied this function from https://github.com/mozamimy/chromium-updater modified it by hand. Sorry, I didn't know how else to do it. 
So Thanks @mozamimy