Simple Daemon
Created a simple daemon that runs the script, minus most of the output. The script checks for updates. If there is a new version of chromium it alerts via Apple Script to the User Interface. Press <OK> to quit and install the update. 
You can install the daemon with cron or Launchd (Check http://launchd.info/ for more info on launchd)